Rename drivers after Capybara library
These can't be reused for other frameworks and seeing as how this could
be configurable in the future these should be named after the library,
in this case Capybara.

It's possible in the future a selenium driver could be used with a
library that isn't capybara and this would prevent renaming and
confusion.