Cleanup Rails provided helpers
1. Remove Rails provided assertions for selectors

This has been moved upstream to Capybara so it no longer needs to be
included in Rails.

2. Clean up screenshot helper

Updates documentation to be clearer and separates the concerns of saving
the image, setting the image path, and displaying the image.

3. Remove form helper

The form helper is pretty specific to Basecamp's needs and may not be
helpful outside of Rails.