Fix establishing a connection
When using a multi-db yaml file we don't have a default `:development`
config so this was blowing up. For now, establish a connection with
`:primary`. I want to refactor this later so we don't need to do this
but this is to get past the point it's not working so I can work on
refactoring this for the rake tasks.