WIP - demonstrating Matthew's suggestion for connection switching
The public API is the same as
https://github.com/rails/rails/compare/master...eileencodes:multi-db-improvements-part-4?expand=1

But internally instead of having multiple handlers each holding a pool
list for the connections related to that handler, the
connection_specification_name handles the switching. We have 1 handler
with multiple pool. The write conn would have a
connection_specification_name of "AnimalsModel" and the readonly conn
would be "AnimalsModel.readonly".