Create failing test case for `from` with subquery
After #35360 was merged we bumped Rails again and while the test script
we sent passed we still had the same failures. I debugged a bit and
found that the new code assumes that the `from` is passing only a table
name and that we can get that from `name` or from `value` but there are
cases (like this test case) where you may not pass a name along with the
`FromClause` and the value may not be just a table name.

This test demonstrates what happens when you add an index to the from
clause.