Move pluck up to the association
We call pluck on the association anyway so we can move it up so that we
don't need to call it on the association through the scope.

This will help us implement a feature to split up joins when using
mulitple databases. We will be able to implement a `has_many :through`
that understands how to split up the query into 2 queries instead of
using a join when those records live in different databases.