Pass `role` through connection methods
This change passes `role` around to all the connection methods. At the
moment this `role` isn't being used by the connection methods because
we're still using the `handler` keys as the `role`. Eventually the
`role` used by applications will be this `role` rather than the `role`
on the handler and we'll have a single handler.