Add test for `has_one` reload strict loading bug
If you have a `has_one` relationship that is eager loaded and reload the
relation, strict loading will raise an error even though it's not
causing an N+1.

This only occurs on single relationships like `developer.ship` after
`developer.reload` is called but doesn't happen for
`developers.map(&:ship)` if `developers.reload` is called.