Make sure primary_abstract_class is eager loaded
In #41258 we exposed a public API for changing the default application
record - ie the class used to connect to the database on boot in a
non-eager loaded environment. After attempting to implement this
in GitHub we realized that this won't work unless the class that is the
primary class is eager loaded. Instead of marking a class as the
`primary_abstract_class` this moves it to a config setting. The config
setting must use a string because the class is not loaded yet. When the
database connects it will be constantized. Without this change
application envs that don't eager load like development would see an
error that ApplicationRecord doesn't exist - because the custom
application record class is not loaded yet.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>